requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
python-multipart
slowapi
jinja2
aiofiles